requirements.txt
asgiref==3.11.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
fonttools==4.54.1
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.8
numpy==2.1.3
packaging==24.2
pillow==11.0.0
plyer==2.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.16.0
sqlparse==0.5.5
tzdata==2025.3
Werkzeug==3.1.3
